Move shared seek logic into base, clear fragCurrent on seek out of buffered range
