+++ tests/unit/controller/stream-controller.js
@@ -3,7 +3,8 @@ import sinon from 'sinon';
 import Hls from '../../../src/hls';
 import Event from '../../../src/events';
 import { FragmentTracker, FragmentState } from '../../../src/controller/fragment-tracker';
-import StreamController, { State } from '../../../src/controller/stream-controller';
+import StreamController from '../../../src/controller/stream-controller';
+import { State } from '../../../src/controller/base-stream-controller';
 import M3U8Parser from '../../../src/loader/m3u8-parser';
 import { mockFragments } from '../../mocks/data';
 import Fragment from '../../../src/loader/fragment';
